1. extract duplicate codes of conv, fix lstm/dropout 2. revise test cases
