+++ ngraph/frontend/paddlepaddle/src/op/dropout.cpp
@@ -16,6 +16,9 @@ namespace ngraph
                 NamedOutputs dropout(const NodeContext& node)
                 {
                     auto data = node.get_ng_input("X");
+                    PDPD_ASSERT(node.get_attribute<std::string>("dropout_implementation") ==
+                                    "downgrade_in_infer",
+                                "dropout only supports downgrade_in_infer");
                     auto dropout_prob = ngraph::opset6::Constant::create(
                         ngraph::element::f32, {1}, {1 - node.get_attribute<float>("dropout_prob")});
                     return node.default_single_output_mapping(
