+++ ngraph/test/files/paddlepaddle/gen_scripts/generate_rnn_lstm.py
@@ -8,20 +8,11 @@ def pdpd_rnn_lstm(input_size, hidden_size, layers, direction):
     pdpd.enable_static()
     main_program = pdpd.static.Program()
     startup_program = pdpd.static.Program()
-    w = np.ones((4  * hidden_size, input_size)).astype(np.float32)
-    r = np.ones((4 * hidden_size, hidden_size)).astype(np.float32)
-    b = np.zeros((4 * hidden_size)).astype(np.float32)
+
     num_of_directions = 1 if direction == 'forward' else 2
     with pdpd.static.program_guard(main_program, startup_program):
-        weight_ih = pdpd.ParamAttr(name="weight_ih", initializer=pdpd.nn.initializer.Assign(w))
-        weight_hh = pdpd.ParamAttr(name="weight_hh", initializer=pdpd.nn.initializer.Assign(r))
-        bias_ih_attr = pdpd.ParamAttr(name="bias_ih_attr", initializer=pdpd.nn.initializer.Assign(b))
-        bias_hh_attr = pdpd.ParamAttr(name="bias_hh_attr", initializer=pdpd.nn.initializer.Assign(b))
-        rnn = pdpd.nn.LSTM(input_size, hidden_size, layers, direction,
-                           weight_ih_attr=weight_ih,
-                           weight_hh_attr=weight_hh,
-                           bias_ih_attr=bias_ih_attr,
-                           bias_hh_attr=bias_hh_attr)
+
+        rnn = pdpd.nn.LSTM(input_size, hidden_size, layers, direction)
 
         data = pdpd.static.data(name='x', shape=[4, 3, input_size], dtype='float32')
         prev_h = pdpd.ones(shape=[layers * num_of_directions, 4, hidden_size], dtype=np.float32)
@@ -33,7 +24,6 @@ def pdpd_rnn_lstm(input_size, hidden_size, layers, direction):
         exe = pdpd.static.Executor(cpu[0])
         exe.run(startup_program)
 
-        pdpd.static.io.save_inference_model("../models/paddle_rnn_lstm_layer_" + str(layers) + '_' + str(direction), [data], [y], exe)
         outs = exe.run(
             feed={'x': np.ones([4, 3, input_size]).astype(np.float32)},
             fetch_list=[y],
@@ -52,6 +42,24 @@ if __name__ == "__main__":
             'hidden_size': 2,
             'layers': 1,
             'direction': 'forward',
+        },
+        {
+            'input_size': 2,
+            'hidden_size': 2,
+            'layers': 1,
+            'direction': 'bidirectional',
+        },
+        {
+            'input_size': 2,
+            'hidden_size': 2,
+            'layers': 2,
+            'direction': 'forward',
+        },
+        {
+            'input_size': 2,
+            'hidden_size': 2,
+            'layers': 2,
+            'direction': 'bidirectional',
         }
     ]
