+++ ngraph/test/files/paddlepaddle/models/models.csv
@@ -56,7 +56,10 @@ nearest_upsample_false_0,
 pad3d_test1,
 pad3d_test2,
 pad3d_test3,
+rnn_lstm_layer_1_bidirectional,
 rnn_lstm_layer_1_forward,
+rnn_lstm_layer_2_bidirectional,
+rnn_lstm_layer_2_forward,
 slice,
 slice_1d,
 split_test1,
